requirements.txt
flask
langchain
langchain-community
python-dotenv
requests
gunicorn